requirements.txt
# 屏幕捕获与OCR
pyautogui==0.9.54
opencv-python==4.8.0.76
pytesseract==0.3.10
pillow==10.0.0

# 知识库管理
langchain==0.0.267
langchain_community==0.0.13
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.4.2

# 文档处理
pypdf==3.17.1
python-docx==0.8.11
markdown==3.5

# LLM接口
openrouter==0.5.0
aiohttp==3.8.5

# 工具
python-dotenv==1.0.0
tqdm==4.66.1 